Remove frontpageview, use existing index view
