+++ controlpanel/frontend/jinja2/justice_email.html
@@ -17,16 +17,12 @@
           <button type="submit" class="govuk-button" data-module="govuk-button">
             Authenticate with Justice identity
           </button>
-          <a class="govuk-button govuk-button--secondary" href="/">
+          <a class="govuk-button govuk-button--secondary" href="{{ url('list-tools') }}">
             Skip for now
           </a>
         </form>
       </div>
     </div>
   </div>
-
-
-
-
 
 {% endblock %}
