+++ controlpanel/frontend/views/auth.py
@@ -1,43 +1,15 @@
 # Standard library
-import base64
-import hashlib
 
 # Third-party
 import sentry_sdk
-from authlib.common.security import generate_token
 from authlib.integrations.django_client import OAuthError
 from django.contrib import messages
 from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.views import View
-from django.views.generic import TemplateView
 
 # First-party/Local
 from controlpanel.oidc import OIDCLoginRequiredMixin, oauth
-
-
-class FrontPageView(OIDCLoginRequiredMixin, TemplateView):
-    http_method_names = ["get", "post"]
-    template_name = "frontpage.html"
-
-    def get(self, request, *args, **kwargs):
-        if self.request.user.justice_email:
-            return HttpResponseRedirect(reverse("index"))
-        return super().get(request, *args, **kwargs)
-
-    def post(self, request):
-        code_challenge = self._get_code_challenge()
-        redirect_uri = request.build_absolute_uri(reverse("entraid-auth"))
-        return oauth.azure.authorize_redirect(
-            request,
-            redirect_uri,
-            code_challenge=code_challenge,
-        )
-
-    def _get_code_challenge(self):
-        code_verifier = generate_token(64)
-        digest = hashlib.sha256(code_verifier.encode()).digest()
-        return base64.urlsafe_b64encode(digest).rstrip(b"=").decode()
 
 
 class EntraIdAuthView(OIDCLoginRequiredMixin, View):
